Created a new preference class called FilePreference
File preference is a SwitchPreference kinda preference. When switched on it creates file (the file name is the key of the preference) and when switched off it removes the file.

File location is in filesDir (/data/data/packagename/files/)

If a developer uses those files for boolean value in some smali, we can create them on boot using boot observer.