Fixed ColorPickerPreference to take hex as defaultValue
There was a generic bug (for ages) in ColorPickerPreference. It could take default value as reference to resource color, f.e. "@color/myColorWhite" and then it would persist 0 into shared prefrences. Or it would take an actual integer for color. If you tried to put hex value it would crash the app.

Fixed it in the original source for the ColorPickerPreference by changing the protected Object onGetDefaultValue method. Now it can take integer or hex string consisting of full ARGB or RGB with 6 or 8 chars after hash (#).